+++ NEWS.md
@@ -6,6 +6,7 @@
   - Rename S4-method `uexIntensities()` to `calcExShockArrivalIntensities()`
   - Rename S4-method `exQMatrix()` to `calcMDCMGeneratorMatrix()`
   - Rename S4-method `intensities()` to `calcShockArrivalIntensities()`
+  - Rename S4-method `defaultMethod()` to `getDefaultMethodString()`
 
 - Improve extensibility by exposing `defaultMethod` and `valueOf0` (#118)